requirements.txt
tk
Pillow
opencv-python
numpy
ultralytics
